First version of the SPI driver BCM2835 / BCM2711 controllers
The driver provides a factory interface to create new SPI devices with
an RPC interface. If the SPI device is present on the SPI bus, the
create call at the factory succeeds and the client can then interact
with the just created device via its RPC interface (see
spi_device_if.h).

Change-Id: Ieb5baead82ca6c9887725122b3e11368290c032b
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>