Add support for VirtIO SPI controller type
This adds another device type to create at the factory. The first
parameter in the Factory.create() call defines the device type.
Type 0 creates a virtio-spi-controller,
Type 1 creates a RPC device.

The virtio-spi-controller type can be passed to uvmm via a
virtio-proxy device device in the uvmm DT, such that a linux guest can
make use of the SPI device.

Requires-Change-Id: I03d8d0eb7a9291620e274b82bb16b105e4e1550c
Change-Id: Idcd7f1abe16e0cb83bcd7e1666e58a8335e10442
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>